Refactoring: sorting steps to  pages objects (make new classes <page>Steps)
