+++ src/test/java/runners/StartTests.java
@@ -8,8 +8,8 @@ import org.junit.runner.RunWith;
 @RunWith(Cucumber.class)
 @CucumberOptions(
         features = "src/test/resources/features/account_management_smoke_test.feature",
-        glue = {"/tests"},
-        plugin = {"pretty", "html:cucumber-html-report/cucumber-report.html","json:cucumber-json-report/cucumber.json"},
+        glue = {"/steps"},
+        plugin = {"html:cucumber-html-report/cucumber-report.html", "json:cucumber-json-report/cucumber.json", "pretty"},
         snippets = CucumberOptions.SnippetType.CAMELCASE
 
 )
