+++ src/test/java/steps/CreateAccountMessageSteps.java
@@ -1,4 +1,4 @@
-package tests;
+package steps;
 
 public class CreateAccountMessageSteps {
 }
