+++ src/test/java/steps/DeleteAccountMessageSteps.java
@@ -1,4 +1,4 @@
-package tests;
+package steps;
 
 public class DeleteAccountMessageSteps {
 }
