+++ src/test/java/steps/DeleteAccountSteps.java
@@ -1,4 +1,4 @@
-package tests;
+package steps;
 
 
 import org.junit.jupiter.api.*;
