+++ src/test/java/steps/HomePageSteps.java
@@ -1,0 +1,7 @@
+package steps;
+
+public class HomePageSteps extends BaseStep{
+    public void navigateToHomePage() {
+        getDriver().get(getBASE_URL());
+    }
+}
