+++ src/test/java/steps/LogoutSteps.java
@@ -1,4 +1,4 @@
-package tests;
+package steps;
 
 public class LogoutSteps extends BaseStep {
     public void logoutLinkInNavBarTest() {
