--- src/test/java/tests/HomePageSteps.java
@@ -1,4 +1,0 @@
-package tests;
-
-public class HomePageSteps {
-}
