--- src/test/java/tests/LoginPageSteps.java
@@ -1,4 +1,0 @@
-package tests;
-
-public class LoginPageSteps {
-}
