--- src/test/java/tests/NavigationBarSteps.java
@@ -1,4 +1,0 @@
-package tests;
-
-public class NavigationBarSteps {
-}
