--- src/test/java/tests/SignupPageSteps.java
@@ -1,4 +1,0 @@
-package tests;
-
-public class SignupPageSteps {
-}
